Extract functions to utils
